requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
awsebcli==3.20.3
backports.zoneinfo==0.2.1
botocore==1.23.54
cachetools==5.0.0
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.3
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flatbuffers==2.0
future==0.16.0
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
jmespath==0.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
networkx==2.6.3
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
protobuf==3.19.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
pysqlite3==0.4.7
pysqlite3-binary==0.4.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.7.3
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
tensorflow==2.8.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2021.11.2
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.1.9
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
